feat: implementation of private routes for the administrator and user. Routes depending on the user's role
