+++ src/Layout/LayoutAuthorized/LayoutAuthorized.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Outlet } from "react-router-dom";
+
+export default function LayoutMain() {
+  return <Outlet />;
+}
