+++ src/Layout/LayoutMain/LayoutMain.css
@@ -1,15 +1,3 @@
-@import url("https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;500;600;700;800&display=swap");
-
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: "Open Sans", sans-serif;
-}
-
 .App {
   min-height: 100vh;
   display: flex;
