+++ src/Roles/Roles.jsx
@@ -1,0 +1,11 @@
+const ROLES = {
+  User: {
+    clave: 2001,
+    url: "user"
+  },
+  Admin: {
+    clave: 5150,
+    url: "admin"
+  }
+};
+export default ROLES
